requirements.txt
numpy==2.3.5
sympy==1.14.0
pulp==3.3.0
pillow==12.0.0
matplotlib==3.10.8
